+++ NEWS.md
@@ -34,6 +34,9 @@ from a BDS dataset, e.g. ADLB. (#943)
 - `derive_param_wbc_abs()` - Adds a parameter for lab differentials converted to absolute values. 
 (#941)
 
+- New function `derive_param_first_event()`: Add a new parameter for the first
+event occurring within a parameter of the input dataset. (#1063)
+
 - New high order functions (#701):
 
   - `restrict_derivation()` - Execute a derivation on a subset of the input dataset